Revert "feat: support canary test for multiple load balancers (#80)"
This reverts commit 681eccf23d886344e5c0bafb8420739ff63ff41c.
