--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-  "explorer.fileNesting.enabled": true,
-  "explorer.fileNesting.patterns": {
-    "*.go": "${capture}_test.go",
-    "go.mod": "go.sum"
-  }
-}
